+++ CHANGELOG.md
@@ -2,13 +2,15 @@
 
 ## [1.2.0] - 2024-04-03
 ### 🚀 New Features
-- Add yml configuration `show_add` and `show_edit` to `task-organizer-card`
+- Added yml configuration `show_add` and `show_edit` to `task-organizer-card`
 - Possibility to set a specific due date
+- Added area for task
 
 ### 🐛 Bug Fixes
 - All tooltips in the cards have been translated
 - All cards now have a customizable title
 - Sensor `task_organizer_due_tasks` transfered into new sensor `task_organizer_due_and_overdue_tasks` and new sensors for differenciation  `task_organizer_due_tasks` and `task_organizer_overdue_tasks` (without paused tasks)
+- Design adapted to home assistant
 
 ### ⚠️ Breaking Changes
 No breaking changes